third_party/blink/web_tests/platform/mac-mac12/external/wpt/html/semantics/forms/the-textarea-element/wrapping-transformation.window-expected.txt

This is a testharness.js-based test.
[FAIL] Textarea hard-wrapping should honor the col count unconditionally,
  assert_equals: expected "ABCDEFGHIJ\\nKLMNOPQRST\\nUVWXYZ" but got "ABCDEFGHIJKL\\nMNOPQRSTUVWX\\nYZ"
Harness: the test ran to completion.
